# docs/swagger.yaml
definitions:
  response.GetUser:
    properties:
      id:
        type: integer
      name:
        description: 用户名
        type: string
      phone:
        description: 手机号
        type: string
    type: object
  response.GetUserList:
    properties:
      id:
        type: integer
      name:
        description: 用户名
        type: string
      phone:
        description: 手机号
        type: string
    type: object
  response.Login:
    properties:
      expires_at:
        type: integer
      token:
        type: string
      userInfo:
        $ref: '#/definitions/response.GetUser'
    type: object
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: gin-exampl api.
  title: gin-example api
  version: 1.0.0
paths:
  /login:
    post:
      consumes:
      - multipart/form-data
      description: 用户登录
      parameters:
      - description: 用户名
        in: formData
        name: name
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.Login'
              type: object
      summary: 用户登录
      tags:
      - 公共
  /logout:
    get:
      consumes:
      - multipart/form-data
      description: 登出
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 登出
      tags:
      - 用户
  /register:
    post:
      consumes:
      - multipart/form-data
      description: 用户注册
      parameters:
      - description: 用户名
        in: formData
        name: name
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 手机号
        in: formData
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 公共
  /user/{id}:
    get:
      consumes:
      - multipart/form-data
      description: 获取用户
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.GetUser'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户
      tags:
      - 用户
  /user/GetUserList:
    get:
      consumes:
      - multipart/form-data
      description: 获取用户列表
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.GetUserList'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - 用户
  /user/UpdateUser:
    post:
      consumes:
      - multipart/form-data
      description: 更新用户
      parameters:
      - description: 用户名
        in: formData
        name: name
        required: true
        type: string
      - description: 手机号
        in: formData
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - 用户
  /user/changePassword:
    post:
      consumes:
      - multipart/form-data
      description: 修改密码
      parameters:
      - description: 新密码
        in: formData
        name: newPassword
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 用户
  /user/deleteUser:
    post:
      consumes:
      - multipart/form-data
      description: 删除用户
      parameters:
      - in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"